clipboard/requirements.txt
aniso8601==2.0.0
boto3==1.7.8
botocore==1.10.8
click==6.7
docutils==0.14
Flask==0.12.2
flask-marshmallow==0.8.0
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-StatsD==0.1.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
pbr==3.1.1
Pillow==5.1.0
pip==9.0.1
python-dateutil==2.7.2
pytz==2018.3
s3transfer==0.1.13
setuptools==28.8.0
six==1.11.0
SQLAlchemy==1.2.2
statsd==3.2.2
stevedore==1.28.0
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.7.2
Werkzeug==0.14.1
wheel==0.30.0
